+++ validator/src/main/java/org/grnet/status/validators/EventStatusValidator.java
@@ -3,9 +3,7 @@ package org.grnet.status.validators;
 import io.quarkus.runtime.annotations.RegisterForReflection;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
-import org.grnet.status.constraints.ValidEventName;
 import org.grnet.status.constraints.ValidEventStatus;
-import org.grnet.status.enums.ContactType;
 import org.grnet.status.enums.EventStatus;
 
 import java.util.Arrays;